feature/fsel-3838: Update Resolve config comment

+++ Services/Fsel.Course/Fsel.Course.Domain/Models/EntityModels/OverallReportPlacementTestModel.cs
@@ -4,14 +4,14 @@ namespace Fsel.Course.Domain.Models.EntityModels
 {
     public class OverallReportPlacementTestModel
     {
-        public double TotalStudent { get; set; }
-        public double TotalPlacementTest { get; set; }
-        public double TotalCompletePlacementTest { get; set; }
-        public double TotalCourseLevelA1 { get; set; }
-        public double TotalCourseLevelA2 { get; set; }
-        public double TotalCourseLevelB1 { get; set; }
-        public double TotalCourseLevelB1Plus { get; set; }
-        public double TotalCourseLevelB2 { get; set; }
-        public double TotalCourseLevelC1 { get; set; }
+        public long TotalStudent { get; set; }
+        public long TotalPlacementTest { get; set; }
+        public long TotalCompletePlacementTest { get; set; }
+        public long TotalCourseLevelA1 { get; set; }
+        public long TotalCourseLevelA2 { get; set; }
+        public long TotalCourseLevelB1 { get; set; }
+        public long TotalCourseLevelB1Plus { get; set; }
+        public long TotalCourseLevelB2 { get; set; }
+        public long TotalCourseLevelC1 { get; set; }
     }
 }
